+++ Editor/Drawers/ButtonDrawers/DecButtonDrawer/DecButtonAttributeDrawerUIToolkit.cs
@@ -15,6 +15,10 @@ using UnityEditor;
 using UnityEngine;
 using UnityEngine.UIElements;
 
+#if SAINTSFIELD_UNITASK && !SAINTSFIELD_UNITASK_DISABLE
+using Cysharp.Threading.Tasks;
+#endif
+
 namespace SaintsField.Editor.Drawers.ButtonDrawers.DecButtonDrawer
 {
     public partial class DecButtonAttributeDrawer
@@ -99,6 +103,9 @@ namespace SaintsField.Editor.Drawers.ButtonDrawers.DecButtonDrawer
                     new Dictionary<Waiter, (MethodInfo methodInfo, object parent)>();
                 foreach ((MethodInfo methodInfo, object result) in results)
                 {
+#if SAINTSFIELD_UNITASK && !SAINTSFIELD_UNITASK_DISABLE
+                    (bool returnIsUniTask, Type returnUniTaskValueType) = GetUniTaskReturnInfo(methodInfo.ReturnType);
+#endif
                     if (result is IEnumerator ie)
                     {
                         buttonUserData.Enumerators.Add(new Waiter(ie));
@@ -113,6 +120,22 @@ namespace SaintsField.Editor.Drawers.ButtonDrawers.DecButtonDrawer
                             waiterReturnTargets[waiter] = (methodInfo, refreshedParent);
                         }
                     }
+#if SAINTSFIELD_UNITASK && !SAINTSFIELD_UNITASK_DISABLE
+                    else if (result is UniTask uniTask)
+                    {
+                        buttonUserData.Enumerators.Add(new Waiter(uniTask));
+                    }
+                    else if (returnIsUniTask && returnUniTaskValueType != null)
+                    {
+                        Waiter waiter = Waiter.UniTaskWithValue(result, returnUniTaskValueType);
+                        buttonUserData.Enumerators.Add(waiter);
+                        if (!decButtonAttribute.HideReturnValue)
+                        {
+                            refreshedParent ??= SerializedUtils.GetFieldInfoAndDirectParent(property).parent;
+                            waiterReturnTargets[waiter] = (methodInfo, refreshedParent);
+                        }
+                    }
+#endif
                     else if (!decButtonAttribute.HideReturnValue)
                     {
                         if(result != null && result.GetType() != typeof(void))